Struct varisat_checker::context::Context

pub struct Context<'a> {
    pub checker_state: CheckerState,
    pub clause_hasher: ClauseHasher,
    pub clauses: Clauses,
    pub processing: Processing<'a>,
    pub rup_check: RupCheck,
    pub tmp_data: TmpData,
    pub variables: Variables,
}

Central checker data structure.

This struct contains all data kept by the checker. Most functions operating on multiple fields of the context use partial references provided by the partial_ref crate. This documents the data dependencies and makes the borrow checker happy without the overhead of passing individual references.

Fields
checker_state: CheckerStateclause_hasher: ClauseHasherclauses: Clausesprocessing: Processing<'a>rup_check: RupChecktmp_data: TmpDatavariables: Variables