Struct varisat_checker::hash::ClauseHasher

pub struct ClauseHasher {
    pub hash_bits: u32,
    pub buffered_solver_var_names: Vec<(Var, Option<Var>)>,
    pub rename_in_buffered_solver_var_names: bool,
    solver_var_names: HashMap<Var, Var>,
}

Fields
hash_bits: u32

How many bits are used for storing clause hashes.

buffered_solver_var_names: Vec<(Var, Option<Var>)>

Changed solver names that are not yet reflected in the checkers current clause hashes.

rename_in_buffered_solver_var_names: bool

Does buffered_solver_var_names contain a new name?

If it contains only deletes, there is no need to rehash

solver_var_names: HashMap<Var, Var>

Current mapping from global var names to solver var names, used for hashing.

Methods

impl ClauseHasher

pub fn clause_hash(&self, lits: &[Lit]) -> ClauseHash

Compute a clause hash of the current bit size